Move types to its own folder
