+++ packages/useWorker/src/lib/createWorkerBlobUrl.ts
@@ -1,5 +1,5 @@
 // import isoworker from 'isoworker'
-import { TRANSFERABLE_TYPE } from '../useWorker'
+import { TRANSFERABLE_TYPE } from '../types'
 import jobRunner from './jobRunner'
 import remoteDepsParser from './remoteDepsParser'
